move util code to cpu_adam
